pyproject.toml
[tool.poetry]
name = "photosearch"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
keras = "^2.4.3"
tensorflow = "^2.5.0"
Pillow = "^8.2.0"
starlette = "^0.14.2"
uvicorn = "^0.14.0"
Jinja2 = "^3.0.1"
aiofiles = "^0.7.0"
matplotlib = "^3.4.2"
mtcnn = "^0.1.0"
keras-vggface = "^0.6"
Keras-Applications = "^1.0.8"
scikit-learn = "^0.24.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pynvim = "^0.4.3"
jupyter = "^1.0.0"
notebook = "^6.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
